requirements.txt
streamlit 
langchain 
python-dotenv
cloudinary 
open-interpreter 
requests